# docs/openapi-main.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/orders/calculate":{"post":{"summary":"Calculate Order","operationId":"calculate_order_orders_calculate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCalculateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/establishments":{"get":{"summary":"Get Establishments Endpoint","operationId":"get_establishments_endpoint_establishments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/EstablishmentResponse"},"type":"array","title":"Response Get Establishments Endpoint Establishments Get"}}}}}},"post":{"summary":"Create Establishments Endpoint","operationId":"create_establishments_endpoint_establishments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstablishmentCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstablishmentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/establishments/{establishment_id}/products":{"post":{"summary":"Create Product Endpoint","operationId":"create_product_endpoint_establishments__establishment_id__products_post","parameters":[{"name":"establishment_id","in":"path","required":true,"schema":{"type":"integer","title":"Establishment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductResponse"}}}},"404":{"description":"Establishment not found"},"409":{"description":"Product already exists"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Products Endpoint","operationId":"get_products_endpoint_establishments__establishment_id__products_get","parameters":[{"name":"establishment_id","in":"path","required":true,"schema":{"type":"integer","title":"Establishment Id"}},{"name":"only_available","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Only Available"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"},"title":"Response Get Products Endpoint Establishments  Establishment Id  Products Get"}}}},"404":{"description":"Establishment not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/establishments/{establishment_id}/orders":{"post":{"summary":"Create Order Endpoint","operationId":"create_order_endpoint_establishments__establishment_id__orders_post","parameters":[{"name":"establishment_id","in":"path","required":true,"schema":{"type":"integer","title":"Establishment Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders":{"get":{"summary":"Get Orders Endpoint","operationId":"get_orders_endpoint_orders_get","parameters":[{"name":"establishment_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Establishment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"},"title":"Response Get Orders Endpoint Orders Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}":{"get":{"summary":"Get Order Endpoint","operationId":"get_order_endpoint_orders__order_id__get","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}/status":{"patch":{"summary":"Update Order Status Endpoint","operationId":"update_order_status_endpoint_orders__order_id__status_patch","parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"integer","title":"Order Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderStatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EstablishmentCreate":{"properties":{"external_id":{"type":"string","maxLength":100,"minLength":1,"title":"External Id"},"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"establishment_type":{"type":"string","maxLength":30,"minLength":1,"title":"Establishment Type"},"address":{"type":"string","maxLength":500,"minLength":1,"title":"Address"}},"type":"object","required":["external_id","name","establishment_type","address"],"title":"EstablishmentCreate"},"EstablishmentResponse":{"properties":{"external_id":{"type":"string","maxLength":100,"minLength":1,"title":"External Id"},"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"establishment_type":{"type":"string","maxLength":30,"minLength":1,"title":"Establishment Type"},"address":{"type":"string","maxLength":500,"minLength":1,"title":"Address"},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["external_id","name","establishment_type","address","id","is_active"],"title":"EstablishmentResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrderCalculateRequest":{"properties":{"items":{"items":{"$ref":"#/components/schemas/OrderItem"},"type":"array","minItems":1,"title":"Items"},"delivery_price":{"type":"number","minimum":0.0,"title":"Delivery Price"},"discount":{"type":"number","minimum":0.0,"title":"Discount"}},"type":"object","required":["items","delivery_price","discount"],"title":"OrderCalculateRequest"},"OrderCreate":{"properties":{"customer_name":{"type":"string","maxLength":100,"minLength":1,"title":"Customer Name"},"customer_phone":{"type":"string","maxLength":30,"minLength":5,"title":"Customer Phone"},"delivery_address":{"type":"string","maxLength":500,"minLength":5,"title":"Delivery Address"},"items":{"items":{"$ref":"#/components/schemas/OrderItemCreate"},"type":"array","minItems":1,"title":"Items"}},"type":"object","required":["customer_name","customer_phone","delivery_address","items"],"title":"OrderCreate"},"OrderItem":{"properties":{"name":{"type":"string","minLength":1,"title":"Name"},"price":{"type":"number","exclusiveMinimum":0.0,"title":"Price"},"quantity":{"type":"integer","exclusiveMinimum":0.0,"title":"Quantity"}},"type":"object","required":["name","price","quantity"],"title":"OrderItem"},"OrderItemCreate":{"properties":{"product_id":{"type":"integer","exclusiveMinimum":0.0,"title":"Product Id"},"quantity":{"type":"integer","exclusiveMinimum":0.0,"title":"Quantity"},"expected_price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"}],"title":"Expected Price"}},"type":"object","required":["product_id","quantity","expected_price"],"title":"OrderItemCreate"},"OrderItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"product_id":{"type":"integer","title":"Product Id"},"external_product_id":{"type":"string","title":"External Product Id"},"product_name":{"type":"string","title":"Product Name"},"unit_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Unit Price"},"quantity":{"type":"integer","title":"Quantity"},"line_total":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Line Total"}},"type":"object","required":["id","product_id","external_product_id","product_name","unit_price","quantity","line_total"],"title":"OrderItemResponse"},"OrderResponse":{"properties":{"id":{"type":"integer","title":"Id"},"establishment_id":{"type":"integer","title":"Establishment Id"},"customer_name":{"type":"string","title":"Customer Name"},"customer_phone":{"type":"string","title":"Customer Phone"},"delivery_address":{"type":"string","title":"Delivery Address"},"status":{"type":"string","title":"Status"},"rejection_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rejection Reason"},"items_total":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Items Total"},"delivery_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Delivery Price"},"total_price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Total Price"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"items":{"items":{"$ref":"#/components/schemas/OrderItemResponse"},"type":"array","title":"Items"}},"type":"object","required":["id","establishment_id","customer_name","customer_phone","delivery_address","status","rejection_reason","items_total","delivery_price","total_price","created_at","updated_at","items"],"title":"OrderResponse"},"OrderStatusUpdate":{"properties":{"status":{"type":"string","enum":["pending","accepted","preparing","ready","in_delivery","delivered","rejected","cancelled"],"title":"Status"},"reason":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Reason"}},"type":"object","required":["status"],"title":"OrderStatusUpdate"},"ProductCreate":{"properties":{"external_id":{"type":"string","maxLength":100,"minLength":1,"title":"External Id"},"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"category":{"type":"string","maxLength":50,"minLength":1,"title":"Category"},"price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)"}],"title":"Price"},"is_available":{"type":"boolean","title":"Is Available","default":true}},"type":"object","required":["external_id","name","category","price"],"title":"ProductCreate"},"ProductResponse":{"properties":{"external_id":{"type":"string","maxLength":100,"minLength":1,"title":"External Id"},"name":{"type":"string","maxLength":200,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Description"},"category":{"type":"string","maxLength":50,"minLength":1,"title":"Category"},"price":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)","title":"Price"},"is_available":{"type":"boolean","title":"Is Available","default":true},"id":{"type":"integer","title":"Id"},"establishment_id":{"type":"integer","title":"Establishment Id"}},"type":"object","required":["external_id","name","category","price","id","establishment_id"],"title":"ProductResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}